+++ src/test/java/org/bnbalint/snackdaddy/controllers/SnackControllerTest.java
@@ -46,10 +46,10 @@ public class SnackControllerTest {
     void test_getAllSnacks_success() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
         Snack snack = new Snack(
                 "Rice Crispie Treat",
@@ -58,7 +58,7 @@ public class SnackControllerTest {
                 2,
                 ingredients
         );
-        snack.setId(1);
+        snack.setId(1L);
         snack.setCreatedAt(DATE);
         snack.setUpdatedAt(DATE);
         System.out.println("Snack = " + snack);
@@ -108,10 +108,10 @@ public class SnackControllerTest {
     void test_addSnack_success() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
 
         // create the one to send in the request
@@ -131,7 +131,7 @@ public class SnackControllerTest {
                 2,
                 ingredients
         );
-        savedSnack.setId(1);
+        savedSnack.setId(1L);
         savedSnack.setCreatedAt(DATE);
         savedSnack.setUpdatedAt(DATE);
 
@@ -161,10 +161,10 @@ public class SnackControllerTest {
     void test_addSnack_conflict() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
         Snack snack = new Snack(
                 "Rice Crispie Treat",
@@ -192,10 +192,10 @@ public class SnackControllerTest {
     void test_addSnack_error() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
         Snack snack = new Snack(
                 "Rice Crispie Treat",
@@ -226,10 +226,10 @@ public class SnackControllerTest {
     void test_updateSnacks_success() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
 
         // create the one to send in the request
@@ -250,7 +250,7 @@ public class SnackControllerTest {
                 2,
                 ingredients
         );
-        savedSnack.setId(1);
+        savedSnack.setId(1L);
         savedSnack.setCreatedAt(DATE);
         savedSnack.setUpdatedAt(DATE);
 
@@ -280,10 +280,10 @@ public class SnackControllerTest {
     void test_updateSnacks_conflict() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
         Snack snack = new Snack(
                 "Rice Crispie Treat",
@@ -312,10 +312,10 @@ public class SnackControllerTest {
     void test_updateSnacks_error() throws Exception {
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
         Snack snack = new Snack(
                 "Rice Crispie Treat",
@@ -348,7 +348,7 @@ public class SnackControllerTest {
      * @param id - the id to set for the ingredient
      * @return the new ingredient, with the id and created/updated fields set
      */
-    Ingredient makeIngredient(String name, int id) {
+    Ingredient makeIngredient(String name, Long id) {
         Ingredient ingredient = new Ingredient(name);
         ingredient.setId(id);
         ingredient.setCreatedAt(DATE);
