+++ src/test/java/org/bnbalint/snackdaddy/models/SnackLogTest.java
@@ -26,10 +26,10 @@ public class SnackLogTest {
 
         //--------------------------------------------------
         // SET VALUES
-        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4);
-        Ingredient ingredient2 = makeIngredient("Margarine", 5);
-        Ingredient ingredient3 = makeIngredient("Marshmallow", 6);
-        Ingredient ingredient4 = makeIngredient("Vanilla", 7);
+        Ingredient ingredient1 = makeIngredient("Rice Crispy Cereal", 4L);
+        Ingredient ingredient2 = makeIngredient("Margarine", 5L);
+        Ingredient ingredient3 = makeIngredient("Marshmallow", 6L);
+        Ingredient ingredient4 = makeIngredient("Vanilla", 7L);
         Ingredient[] ingredients = { ingredient1, ingredient2, ingredient3, ingredient4 };
         Snack snack = new Snack(
                 "Rice Crispie Treat",
@@ -38,7 +38,7 @@ public class SnackLogTest {
                 2,
                 ingredients
         );
-        snack.setId(1);
+        snack.setId(1L);
         snack.setCreatedAt(DATE);
         snack.setUpdatedAt(DATE);
 
@@ -51,12 +51,12 @@ public class SnackLogTest {
                 "#c42323",
                 "logo.com"
         );
-        team.setId(1);
+        team.setId(1L);
         team.setCreatedAt(DATE);
         team.setUpdatedAt(DATE);
 
         SnackLog snackLogEntry = new SnackLog(snack, team, LocalDate.of(2026, Month.JUNE, 1));
-        snackLogEntry.setId(1);
+        snackLogEntry.setId(1L);
         snackLogEntry.setCreatedAt(DATE);
         snackLogEntry.setUpdatedAt(DATE);
         System.out.println("SnackLogEntry = " + snackLogEntry);
@@ -137,7 +137,7 @@ public class SnackLogTest {
      * @param id - the id to set for the ingredient
      * @return the new ingredient, with the id and created/updated fields set
      */
-    Ingredient makeIngredient(String name, int id) {
+    Ingredient makeIngredient(String name, Long id) {
         Ingredient ingredient = new Ingredient(name);
         ingredient.setId(id);
         ingredient.setCreatedAt(DATE);
